webapp/final/requirements.txt
click==8.0.4
Flask==2.1.0
gunicorn==20.1.0
importlib-metadata==4.11.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
mysql-connector-python==8.0.29
protobuf==4.21.1
python-dotenv==0.19.2
Werkzeug==2.0.3
zipp==3.7.0
